refactor(BA-6609): move api-layer primitives consumed by the service layer down to the data layer
Remove service -> api import violations by relocating the shared
primitives to the data layer and re-exporting them from their former
api locations for backward compatibility:

- Undefined/undefined (and the Singleton metaclass) -> data/common/sentinel.py
- ManagerStatus -> data/manager_status/types.py
- BaseFilterAdapter -> data/filter/adapter.py

The three violating service modules now import from the new data
locations; api.utils, api.__init__, and api.rest.adapter re-export the
moved symbols so existing api-layer consumers keep resolving.
