+++ src/ai/backend/manager/data/manager_status/types.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class ManagerStatus(enum.StrEnum):
+    TERMINATED = "terminated"  # deprecated
+    PREPARING = "preparing"  # deprecated
+    RUNNING = "running"
+    FROZEN = "frozen"
