+++ src/ai/backend/manager/services/manager_admin/service.py
@@ -4,7 +4,7 @@ import socket
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from ai.backend.manager.api import ManagerStatus
+from ai.backend.manager.data.manager_status.types import ManagerStatus
 from ai.backend.manager.errors.api import InvalidAPIParameters
 from ai.backend.manager.errors.resource import InstanceNotFound
 from ai.backend.manager.models.health import get_manager_db_cxn_status
